feat: Remove prover db from house keeper
This PR is a follow-up on
https://github.com/matter-labs/zksync-era/pull/2666, namely the remove
prover side from house keeper.

This PR contains:
- remove all prover jobs from house keeper (now in PJM)
- move core metrics from prover jobs to l1 batch metrics reporter
- remove old configuration

With these changes core & prover are fully decoupled. This will enable
removing unnecessary databases across all envs that don't run provers.
Alongside, core and prover deployments are independent.
