+++ core/lib/protobuf_config/src/proto/config/house_keeper.proto
@@ -4,17 +4,17 @@ package zksync.config.house_keeper;
 
 message HouseKeeper {
     optional uint64 l1_batch_metrics_reporting_interval_ms = 1; // required; ms
-    optional uint64 gpu_prover_queue_reporting_interval_ms = 2; // required; ms
-    optional uint64 prover_job_retrying_interval_ms = 3; // required; ms
-    optional uint64 prover_stats_reporting_interval_ms = 4; // required ms
-    optional uint64 witness_job_moving_interval_ms = 5; // required; ms
-    optional uint64 witness_generator_stats_reporting_interval_ms = 6; // required; ms
-    optional uint64 witness_generator_job_retrying_interval_ms = 9; // required; ms
-    optional uint32 prover_db_pool_size = 10; // required
-    optional uint64 proof_compressor_job_retrying_interval_ms = 12; // required; ms
-    optional uint64 proof_compressor_stats_reporting_interval_ms = 13; // required; ms
-    optional uint64 prover_job_archiver_archiving_interval_ms = 14; // optional; ms
-    optional uint64 prover_job_archiver_archive_after_secs = 15; // optional; seconds
-    optional uint64 fri_gpu_prover_archiver_archiving_interval_ms = 16; // optional; ms
-    optional uint64 fri_gpu_prover_archiver_archive_after_secs = 17; // optional; seconds
+    reserved 2; reserved "gpu_prover_queue_reporting_interval_ms";
+    reserved 3; reserved "prover_job_retrying_interval_ms";
+    reserved 4; reserved "prover_stats_reporting_interval_ms";
+    reserved 5; reserved "witness_job_moving_interval_ms";
+    reserved 6; reserved "witness_generator_stats_reporting_interval_ms";
+    reserved 9; reserved "witness_generator_job_retrying_interval_ms";
+    reserved 10; reserved "prover_db_pool_size";
+    reserved 12; reserved "proof_compressor_job_retrying_interval_ms";
+    reserved 13; reserved "proof_compressor_stats_reporting_interval_ms";
+    reserved 14; reserved "prover_job_archiver_archiving_interval_ms";
+    reserved 15; reserved "prover_job_archiver_archive_after_secs";
+    reserved 16; reserved "fri_gpu_prover_archiver_archiving_interval_ms";
+    reserved 17; reserved "fri_gpu_prover_archiver_archive_after_secs";
 }
