+++ core/node/house_keeper/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod blocks_state_reporter;
+mod metrics;
 pub mod periodic_job;
-pub mod prover;
