--- core/node/house_keeper/src/prover/archiver/mod.rs
@@ -1,5 +1,0 @@
-mod fri_gpu_prover_archiver;
-mod fri_prover_jobs_archiver;
-
-pub use fri_gpu_prover_archiver::FriGpuProverArchiver;
-pub use fri_prover_jobs_archiver::FriProverJobsArchiver;
