--- core/node/house_keeper/src/prover/mod.rs
@@ -1,14 +1,0 @@
-mod archiver;
-mod metrics;
-mod queue_reporter;
-mod retry_manager;
-mod waiting_to_queued_fri_witness_job_mover;
-
-pub use archiver::{FriGpuProverArchiver, FriProverJobsArchiver};
-pub use queue_reporter::{
-    FriProofCompressorQueueReporter, FriProverQueueReporter, FriWitnessGeneratorQueueReporter,
-};
-pub use retry_manager::{
-    FriProofCompressorJobRetryManager, FriProverJobRetryManager, FriWitnessGeneratorJobRetryManager,
-};
-pub use waiting_to_queued_fri_witness_job_mover::WaitingToQueuedFriWitnessJobMover;
