--- core/node/house_keeper/src/prover/retry_manager/mod.rs
@@ -1,7 +1,0 @@
-mod fri_proof_compressor_job_retry_manager;
-mod fri_prover_job_retry_manager;
-mod fri_witness_generator_jobs_retry_manager;
-
-pub use fri_proof_compressor_job_retry_manager::FriProofCompressorJobRetryManager;
-pub use fri_prover_job_retry_manager::FriProverJobRetryManager;
-pub use fri_witness_generator_jobs_retry_manager::FriWitnessGeneratorJobRetryManager;
