+++ prover/crates/lib/prover_dal/src/fri_prover_dal.rs
@@ -476,33 +476,6 @@ impl FriProverDal<'_, '_> {
         }
     }
 
-    pub async fn min_unproved_l1_batch_number_for_aggregation_round(
-        &mut self,
-        aggregation_round: AggregationRound,
-    ) -> Option<L1BatchNumber> {
-        sqlx::query!(
-            r#"
-            SELECT
-                l1_batch_number
-            FROM
-                prover_jobs_fri
-            WHERE
-                status <> 'skipped'
-                AND status <> 'successful'
-                AND aggregation_round = $1
-            ORDER BY
-                l1_batch_number ASC
-            LIMIT
-                1
-            "#,
-            aggregation_round as i16
-        )
-        .fetch_optional(self.storage.conn())
-        .await
-        .unwrap()
-        .map(|row| L1BatchNumber(row.l1_batch_number as u32))
-    }
-
     pub async fn update_status(&mut self, id: u32, status: &str) {
         sqlx::query!(
             r#"
